# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Screener API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ScreenRequest {
  /** JSON array of selected roles */
  roles: string;
  /** JSON array of custom keywords */
  customKeywords?: string;
  jobDescriptions?: Blob[];
  companyInfo?: Blob;
  shortlistedResumes?: Blob[];
  resumes: Blob[];
}

export interface CandidateResult {
  name: string;
  email: string;
  phone: string;
  graduation_year: string;
  college: string;
  degree: string;
  skills_found: string[];
  score: number;
  role_fit: string;
  best_role: string;
  suggested_role: string;
  summary: string;
  skill_gap: string;
  resume_link: string;
}

export interface ScreenResponse {
  results: CandidateResult[];
  processedCount: number;
}

export interface DownloadExcelRequest {
  results: CandidateResult[];
}

export interface ErrorResponse {
  error: string;
  details?: string;
}