# lib/api-zod/src/generated/types/candidateResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Screener API
 * OpenAPI spec version: 0.1.0
 */

export interface CandidateResult {
  name: string;
  email: string;
  phone: string;
  graduation_year: string;
  college: string;
  degree: string;
  skills_found: string[];
  score: number;
  role_fit: string;
  best_role: string;
  suggested_role: string;
  summary: string;
  skill_gap: string;
  resume_link: string;
}