# lib/api-zod/src/generated/types/screenRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Screener API
 * OpenAPI spec version: 0.1.0
 */

export interface ScreenRequest {
  /** JSON array of selected roles */
  roles: string;
  /** JSON array of custom keywords */
  customKeywords?: string;
  jobDescriptions?: Blob[];
  companyInfo?: Blob;
  shortlistedResumes?: Blob[];
  resumes: Blob[];
}