# lib/api-zod/src/generated/types/screenResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Screener API
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateResult } from "./candidateResult";

export interface ScreenResponse {
  results: CandidateResult[];
  processedCount: number;
}